refactor: move event handler to new view model for increasing scope
